requirements.txt
#
# This file is autogenerated by pip-compile with python 3.8
# To update, run:
#
#    pip-compile requirements.in
#
asgiref==3.4.1
    # via django
django==3.2.6
    # via
    #   -r requirements.in
    #   django-leaflet
    #   djangorestframework
django-leaflet==0.28.1
    # via -r requirements.in
djangorestframework==3.12.4
    # via
    #   -r requirements.in
    #   djangorestframework-gis
djangorestframework-gis==0.17
    # via -r requirements.in
psycopg2==2.9.1
    # via -r requirements.in
pytz==2021.1
    # via django
sqlparse==0.4.1
    # via django
